+++ src/components/designSystem/Filters/filtersElements/FiltersItemAmount.tsx
@@ -10,7 +10,7 @@ import {
 import { ComboBoxField, TextInputField } from '~/components/form'
 import { useInternationalization } from '~/hooks/core/useInternationalization'
 
-import { FiltersFormValues } from '../FiltersPanelPoper'
+import { FiltersFormValues } from '../FiltersPanelPopper'
 
 type FiltersItemAmountProps = {
   value: FiltersFormValues['filters'][0]['value']
