+++ src/components/designSystem/Filters/filtersElements/FiltersItemInvoiceNumber.tsx
@@ -5,7 +5,7 @@ import { ComboBox } from '~/components/form'
 import { useGetInvoiceNumbersForFilterItemInvoiceNumbersLazyQuery } from '~/generated/graphql'
 import { useInternationalization } from '~/hooks/core/useInternationalization'
 
-import { FiltersFormValues } from '../FiltersPanelPoper'
+import { FiltersFormValues } from '../FiltersPanelPopper'
 
 gql`
   query getInvoiceNumbersForFilterItemInvoiceNumbers($page: Int, $limit: Int, $searchTerm: String) {
